Fix small bugs introduced in CLI restructuring
