+++ vantage6/setup.py
@@ -62,10 +62,11 @@ setup(
     },
     entry_points={
         'console_scripts': [
+            'v6=vantage6.cli.cli:cli_complete',
+            # TODO v5+ remove vnode/vserver/vdev aliases (?)
             'vnode=vantage6.cli.node.cli:cli_node',
             'vserver=vantage6.cli.server.cli:cli_server',
             'vdev=vantage6.cli.dev.cli:cli_dev',
-            'v6=vantage6.cli.cli:cli_complete'
         ]
     }
 )
