+++ vantage6/vantage6/cli/dev/create.py
@@ -19,7 +19,8 @@ from vantage6.common import info, error, generate_apikey
 
 from vantage6.cli.globals import PACKAGE_FOLDER
 from vantage6.cli.context import ServerContext, NodeContext
-from vantage6.cli.server.cli import vserver_import, get_server_context
+from vantage6.cli.server.common import get_server_context
+from vantage6.cli.server.import_ import vserver_import
 from vantage6.cli.utils import prompt_config_name
